Checkin 1.6.21

+++ src/http/httpvhost.cpp
@@ -679,7 +679,7 @@ HttpContext *HttpVHost::bestMatch(const char *pURI, size_t iUriLen)
 {
     HttpContext *pContext = (HttpContext *)m_contexts.bestMatch(pURI, iUriLen);
     LS_DBG_L(ConfigCtx::getCurConfigCtx(), "HttpVHost::bestMatch %.*s (len %d) return %p, loc %s.",
-             iUriLen, pURI, iUriLen, pContext, pContext ? pContext->getLocation() : "nil");
+             (int) iUriLen, pURI, (int) iUriLen, pContext, pContext ? pContext->getLocation() : "nil");
 
     AutoStr2 missURI; //A while URI start with /
     AutoStr2 missLoc;  //A loc should be added to pContext location for the full path
